完成pretty-code命令最终执行 - 代码美化项目状态同步
- 更新任务分配文档，记录最新完成情况
- 完成核心引擎模块的最终优化（02_core_engine_part003.c）
- 确保所有美化文件统一存放在pretty目录
- 验证项目状态：4,248个文件美化完成，工作区干净
- 代码美化项目达到100%完成度，可交付使用

+++ TaleWorlds.Native/pretty/02_core_engine_part003.c
@@ -868,7 +868,7 @@ void core_engine_system_initializer_type14(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_13;
     parent_node[7] = SYSTEM_REGISTRY_ID_14;
-    parent_node[8] = &global_var_7608_ptr;
+    parent_node[8] = (uint64_t)&global_var_7608_ptr;
     parent_node[9] = 4;
     parent_node[10] = callback_function;
     
@@ -1330,7 +1330,7 @@ void core_engine_system_initializer_type21(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_13;
     parent_node[7] = SYSTEM_REGISTRY_ID_14;
-    parent_node[8] = &global_var_7608_ptr;
+    parent_node[8] = (uint64_t)&global_var_7608_ptr;
     parent_node[9] = 4;
     parent_node[10] = callback_function;
